refactor: minor changes for sendSilently method
