+++ app/src/main/java/core/factory/FacadeFactory.java
@@ -8,8 +8,8 @@ import org.telegram.telegrambots.abilitybots.api.util.AbilityExtension;
 import dbdrivers.DbDriver;
 import dbdrivers.factory.DbDriverFactory;
 import dbdrivers.factory.DbDriverFactoryProducer;
-import static models.EnvVars.COMMAND_TYPE;
-import static models.EnvVars.DB_TYPE;
+import static models.system.EnvVars.COMMAND_TYPE;
+import static models.system.EnvVars.DB_TYPE;
 import models.commands.CommandTypes;
 import models.db.DbTypes;
 import telegram.bot.PovaryoshkaBot;
