+++ app/src/main/java/dbdrivers/factory/PostgresDbDriverFactory.java
@@ -2,15 +2,15 @@ package dbdrivers.factory;
 
 import dbdrivers.DbDriver;
 import dbdrivers.postgres.PostgresDbDriver;
-import static models.EnvVars.ALTER_SQL_SCRIPT_PATH;
-import static models.EnvVars.DB_DATABASE;
-import static models.EnvVars.DB_HOST;
-import static models.EnvVars.DB_PASSWORD;
-import static models.EnvVars.DB_PORT;
-import static models.EnvVars.DB_SCHEMA;
-import static models.EnvVars.DB_USERNAME;
-import static models.EnvVars.INIT_SQL_SCRIPT_PATH;
-import static models.EnvVars.IS_DISTRIBUTED_DATABASE;
+import static models.system.EnvVars.ALTER_SQL_SCRIPT_PATH;
+import static models.system.EnvVars.DB_DATABASE;
+import static models.system.EnvVars.DB_HOST;
+import static models.system.EnvVars.DB_PASSWORD;
+import static models.system.EnvVars.DB_PORT;
+import static models.system.EnvVars.DB_SCHEMA;
+import static models.system.EnvVars.DB_USERNAME;
+import static models.system.EnvVars.INIT_SQL_SCRIPT_PATH;
+import static models.system.EnvVars.IS_DISTRIBUTED_DATABASE;
 
 import java.sql.SQLException;
