+++ app/src/main/java/telegram/commands/AbstractCommand.java
@@ -4,8 +4,10 @@ import java.sql.SQLException;
 import java.util.List;
 import java.util.Optional;
 
+import models.commons.SendCommandOption;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jetbrains.annotations.NotNull;
 import org.telegram.telegrambots.abilitybots.api.objects.MessageContext;
 import org.telegram.telegrambots.abilitybots.api.util.AbilityExtension;
 import org.telegram.telegrambots.meta.api.objects.Update;
@@ -32,12 +34,14 @@ public abstract class AbstractCommand implements AbilityExtension {
 
     @NonNull
     protected Optional<Message> sendSilently(@NonNull String message, @NonNull Update update) {
-        return povaryoshkaBot.getSilent().send(message, update.getMessage().getChatId());
+        return sendSilently(message, update, new SendCommandOption(false));
     }
 
     @NonNull
-    protected Optional<Message> sendSilentlyMarkdownOn(@NonNull String message, @NonNull Update update) {
-        return povaryoshkaBot.getSilent().sendMd(message, update.getMessage().getChatId());
+    protected Optional<Message> sendSilently(@NonNull String message, @NonNull Update update, @NotNull SendCommandOption sendCommandOption) {
+        return (sendCommandOption.markdown())
+            ? povaryoshkaBot.getSilent().sendMd(message, update.getMessage().getChatId())
+            : povaryoshkaBot.getSilent().send(message, update.getMessage().getChatId());
     }
 
     @Nullable
@@ -59,4 +63,5 @@ public abstract class AbstractCommand implements AbilityExtension {
         }
         return messageBuilder.toString();
     }
+
 }
